requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
pillow==10.4.0
Pygments==2.18.0
sqlparse==0.5.1
tzdata==2024.1
